// autogpt_platform/frontend/src/app/api/__generated__/models/aPIKeyInfo.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { APIKeyPermission } from "./aPIKeyPermission";
import type { APIKeyInfoExpiresAt } from "./aPIKeyInfoExpiresAt";
import type { APIKeyInfoLastUsedAt } from "./aPIKeyInfoLastUsedAt";
import type { APIKeyInfoRevokedAt } from "./aPIKeyInfoRevokedAt";
import type { APIKeyStatus } from "./aPIKeyStatus";
import type { APIKeyInfoDescription } from "./aPIKeyInfoDescription";

export interface APIKeyInfo {
  user_id: string;
  scopes: APIKeyPermission[];
  type?: "api_key";
  created_at: Date;
  expires_at?: APIKeyInfoExpiresAt;
  last_used_at?: APIKeyInfoLastUsedAt;
  revoked_at?: APIKeyInfoRevokedAt;
  id: string;
  name: string;
  /** The first 8 characters of the key */
  head: string;
  /** The last 8 characters of the key */
  tail: string;
  status: APIKeyStatus;
  description?: APIKeyInfoDescription;
}